app/requirements.txt
streamlit
chromadb
sqlalchemy
mysql-connector-python
sentence-transformers
langchain
langchain-google-genai
tiktoken
python-dotenv
pandas
